+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/useCollaboration.tsx
@@ -6,17 +6,29 @@ import { useBroadcastStore } from '@/stores';
 import { useProviderStore } from '../stores/useProviderStore';
 import { Base64 } from '../types';
 
-export const useCollaboration = (room?: string, initialContent?: Base64) => {
+export const useCollaboration = (
+  room?: string,
+  initialContent?: Base64,
+  canEdit?: boolean,
+) => {
   const collaborationUrl = useCollaborationUrl(room);
   const { setBroadcastProvider } = useBroadcastStore();
   const { provider, createProvider, destroyProvider } = useProviderStore();
 
+  /**
+   * Initialize the provider
+   */
   useEffect(() => {
-    if (!room || !collaborationUrl || provider) {
+    if (!room || !collaborationUrl || provider || canEdit === undefined) {
       return;
     }
 
-    const newProvider = createProvider(collaborationUrl, room, initialContent);
+    const newProvider = createProvider(
+      collaborationUrl,
+      room,
+      canEdit,
+      initialContent,
+    );
     setBroadcastProvider(newProvider);
   }, [
     provider,
@@ -25,6 +37,7 @@ export const useCollaboration = (room?: string, initialContent?: Base64) => {
     initialContent,
     createProvider,
     setBroadcastProvider,
+    canEdit,
   ]);
 
   /**