+++ src/frontend/apps/impress/src/features/docs/doc-management/stores/useProviderStore.tsx
@@ -4,10 +4,13 @@ import { create } from 'zustand';
 
 import { Base64 } from '@/features/docs/doc-management';
 
+import { CollaborationProvider } from '../libs/CollaborationProvider';
+
 export interface UseCollaborationStore {
   createProvider: (
     providerUrl: string,
     storeId: string,
+    canEdit: boolean,
     initialDoc?: Base64,
   ) => HocuspocusProvider;
   destroyProvider: () => void;
@@ -20,7 +23,7 @@ const defaultValues = {
 
 export const useProviderStore = create<UseCollaborationStore>((set, get) => ({
   ...defaultValues,
-  createProvider: (wsUrl, storeId, initialDoc) => {
+  createProvider: (wsUrl, storeId, canEdit, initialDoc) => {
     const doc = new Y.Doc({
       guid: storeId,
     });
@@ -29,10 +32,11 @@ export const useProviderStore = create<UseCollaborationStore>((set, get) => ({
       Y.applyUpdate(doc, Buffer.from(initialDoc, 'base64'));
     }
 
-    const provider = new HocuspocusProvider({
+    const provider = new CollaborationProvider({
       url: wsUrl,
       name: storeId,
       document: doc,
+      canEdit,
     });
 
     set({